Removing transactionIsolationLevel & operations with savepoint from Schema
